requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
matplotlib
Pillow
PyYAML
tqdm
opencv-python
openslide-python
